moved openAI api calls to new file, fixed issues with vector db
